payment_service/requirements.txt
nats-py
fastapi
uvicorn
aiosqlite